Add a test for addTask function.
